Proposed solution to enhanced logging
To assist with diagnostics and defence against rogue queries, we want to
log more details from the request, per #97. This commit creates one
possible solution for consultation: using a JSON formatted message body
to log some of the detailed fields we need in a way that can be easily
read by log-processing software.