Fix Rubocop warnings
Fix a collection of Rubocop warnings, some of which have arisen from
enabling new cops in the `.rubocop.yml`.

In a small number of cases, I have disabled the cop in `.rubocop.yml`,
as it looked too disruptive to implement the suggested solution (or the
resulting gains were not worth the loss of clarity). Specifically
`Lint/DuplicatedBranch` and `Lint/ConstantDefinitionInBlock`.