Add readme instructions for locally API access
This commit updates the README to include instructions on running the
docker container, or the Rails server, locally with `localhost:8080`
mapped to an instance of the data API via an SSH tunnel.

In addition, to make it easier for developers to run the docker
container locally, I have added a `localrun` command to the `Makefile`.