Enable access to Github Package Registry
epimorphics/hmlr-linked-data#89

This commit enables the app to correctly access packages from the
Epimorphics section of the GitHub Package Registry (GPR). This was not
working correctly before, as the the workflow definition was using an
old pattern.

Correcting the workflow to use the shared workflow definition also
entailed ensuring that the `Makefile` target to update the Docker image
was up-to-date and following our current best practice guidelines.