Pull in upstream changes in API logging
epimorphics/hmlr-linked-data#87

This change addresses the request from @andrew-pickin-epi that the
logger should record the API call made, the status, and the duration in
one log entry.

A typical log entry for a call to the API now looks like:

```json
{
  "level": "INFO",
  "ts": "2022-04-04T20:14:58.612Z",
  "rails": {
    "environment": "development",
    "url": "http://localhost:8080/landregistry/id/ppd?_limit=100\u0026search-paon=rose+cottage\u0026searchPath=propertyAddress",
    "message": "GET from API"
  },
  "status": 200,
  "duration": 1.612488262
}
{
  "level": "INFO",
  "ts": "2022-04-04T20:15:03.765Z",
  "rails": {
    "environment": "development",
    "url": "http://localhost:8080/landregistry/id/ppd?_count=%40id\u0026_limit=10000\u0026search-paon=rose+cottage\u0026searchPath=propertyAddress",
    "message": "GET from API"
  },
  "status": 200,
  "duration": 5.1462943
}
```

One consideration is that the times are reported in seconds. In future,
we may want to (well, probably should) standardise on a single units
measurement for the `duration` field. Or at least add a unit.

Note that, en passant, I also fixed a minor bug in the `Dockerfile` that
didn't need its own commit.