fix: keep the message simple silly
Updated the error response to be only the error message itself in production while keeping the rest of the debugging information specific to the development environment. This also reformats the response shown to the end user to also keep it simple while still offering as much info to support tickets as possible. Finally, this also utilises the updated logging levels based on the error status.

https://github.com/epimorphics/hmlr-linked-data/issues/111